requirements.txt
Flask
scikit-learn
numpy
pandas
PyPDF2
python-docx
